geoq/requirements.txt
Django==1.5.1
Django-Select2==4.0.0
Paver==1.2.1
South==0.8.1
django-appconf==0.6
django-bootstrap-toolkit==2.15.0
django-compressor==1.3
django-geoexplorer==3.0.4
django-leaflet==0.6.0a
django-reversion==1.7.1
django-stronghold==0.2.2
django-teamwork==0.0.3
jsonfield==0.9.17
psycopg2==2.5.1
requests==1.2.3
six==1.3.0
wsgiref==0.1.2
